fix code after review
